feat(sdk): Export `FrozenSlidingSyncList` into its own module.
